Swap GUI and WOO (LC, LA, LG) on mac layout; map key for locking screen
This is to smooth the transition between linux and mac. On linux I use
LGUI as a modifier for xmonad, but on mac this is used by the OS so I
use a CTRL, ALT, GUI combo. This was previously mapped to a different
key, so switching between machines meant changing the fingers to
interact with the window manager (amethyst on the mac). This fixes that
issue by swapping GUI for the CTRL, ALT, GUI combo on the mac layout, at
the expense of moving the GUI key on the right keyboard to a less
convenient location.

While making this change I also fixed the macros for the combo keys
(HYPER, MEH, WOO), and added a dedicated key for locking the screen
on the mac.